# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.CreatePostRequest:
    properties:
      body:
        maxLength: 255
        minLength: 5
        type: string
      title:
        maxLength: 255
        minLength: 5
        type: string
    required:
    - body
    - title
    type: object
...
  request.UpdatePostRequest:
    properties:
      body:
        maxLength: 255
        minLength: 5
        type: string
      title:
        description: ID    int    `validate:"required" json:"post_id"`
        maxLength: 255
        minLength: 5
        type: string
    required:
    - body
    - title
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8888
info:
  contact:
    email: rdbhagat999@gmail.com
    name: Ramandeep Bhagat
  description: A demo Service API in Golang using Gin Framework
  termsOfService: https://swagger.io/terms/
  title: Demo Service API
  version: 1.0.0
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Returns a list of post
      operationId: FindAllPost
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all post
    post:
      consumes:
      - application/json
      description: Save post in database
      operationId: CreatePost
      parameters:
      - description: Create post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/request.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create post
  /posts/{postId}:
    delete:
      consumes:
      - application/json
      description: Delete post from database
      operationId: DeletePost
      parameters:
      - description: Delete post by id
        in: path
        name: postId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete post
    get:
      consumes:
      - application/json
      description: Returns a single post when postId maches id
      operationId: FindByIdPost
      parameters:
      - description: Find post by id
        in: path
        name: postId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a single post by its id
    put:
      consumes:
      - application/json
      description: Update and save post in database
      operationId: UpdatePost
      parameters:
      - description: Update post by id
        in: path
        name: postId
        required: true
        type: string
      - description: Update post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update post
  /posts/userposts:
    get:
      consumes:
      - application/json
      description: Returns a list of post
      operationId: FindAllByUserId
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all post by userId
...
swagger: "2.0"